Common: Rename pydantic_event_serializer.py -> pydantic_agent_event_serializer.py
